result viewer layout save load delete rename improved

+++ cravat/webresult/nocache/js/util.js
@@ -600,23 +600,29 @@ function loadLayoutSettingAs (evt) {
     hideAllMenu3();
 	var div = document.getElementById('load_layout_select_div');
 	emptyElement(div);
-    div.style.display = 'block';
+    div.style.display = 'inline-block';
     evt.stopPropagation();
 	$.get('/result/service/getlayoutsavenames', {'username': username, 'job_id': jobId, 'dbpath': dbPath}).done(function (response) {
-        var quickSaveNameIdx = response.indexOf(quickSaveName);
-        if (quickSaveNameIdx >= 0) {
-            response.splice(quickSaveNameIdx, 1);
+        if (response.length == 0) {
+            var a = getEl('a');
+            a.textContent = '(no\xa0saved\xa0layout)';
+            addEl(div, a);
+        } else {
+            var quickSaveNameIdx = response.indexOf(quickSaveName);
+            if (quickSaveNameIdx >= 0) {
+                response.splice(quickSaveNameIdx, 1);
+            }
+            var savedLayoutNames = response;
+            for (var i = 0; i < savedLayoutNames.length; i++) {
+                var name = savedLayoutNames[i];
+                var a = getEl('a');
+                a.textContent = name;
+                a.addEventListener('click', function (evt) {
+                    loadLayoutSetting(evt.target.textContent, null)
+                });
+                addEl(div, a);
+            }
         }
-    	var savedLayoutNames = response;
-    	for (var i = 0; i < savedLayoutNames.length; i++) {
-    		var name = savedLayoutNames[i];
-    		var a = getEl('a');
-    		a.textContent = name;
-    		a.addEventListener('click', function (evt) {
-    			loadLayoutSetting(evt.target.textContent, null)
-    		});
-    		addEl(div, a);
-    	}
     });
 }
 
@@ -664,7 +670,9 @@ function deleteLayoutSettingAs (evt) {
 	var div = document.getElementById('delete_layout_select_div');
 	emptyElement(div);
     div.style.display = 'block';
-    evt.stopPropagation();
+    if (evt != null) {
+        evt.stopPropagation();
+    }
 	$.get('/result/service/getlayoutsavenames', {'username': username, 'job_id': jobId, 'dbpath': dbPath}).done(function (response) {
         /*
         var quickSaveNameIdx = response.indexOf(quickSaveName);
@@ -672,27 +680,34 @@ function deleteLayoutSettingAs (evt) {
             response.splice(quickSaveNameIdx, 1);
         }
         */
-    	savedLayoutNames = response;
-    	for (var i = 0; i < savedLayoutNames.length; i++) {
-    		var name = savedLayoutNames[i];
-    		var a = getEl('a');
-    		a.textContent = name;
-    		a.setAttribute('module', name);
-    		a.addEventListener('click', function (evt) {
-    			var name = evt.target.getAttribute('module');
-    			var yes = confirm('Delete ' + name + '?');
-    			if (yes) {
-    				deleteLayoutSetting(evt.target.textContent, null)
-    			}
-    		});
-    		addEl(div, a);
-    	}
+        if (response.length == 0) {
+            var a = getEl('a');
+            a.textContent = '(no\xa0saved\xa0layout)';
+            addEl(div, a);
+        } else {
+            savedLayoutNames = response;
+            for (var i = 0; i < savedLayoutNames.length; i++) {
+                var name = savedLayoutNames[i];
+                var a = getEl('a');
+                a.textContent = name;
+                a.setAttribute('module', name);
+                a.addEventListener('click', function (evt) {
+                    var name = evt.target.getAttribute('module');
+                    var yes = confirm('Delete ' + name + '?');
+                    if (yes) {
+                        deleteLayoutSetting(evt.target.textContent, null)
+                    }
+                });
+                addEl(div, a);
+            }
+        }
     });
 }
 
 function deleteLayoutSetting (name, callback) {
 	$.get('/result/service/deletelayoutsetting', {'username': username, 'job_id': jobId, 'dbpath': dbPath, 'name': name}).done(function (response) {
 		writeLogDiv('Layout setting deleted');
+        deleteLayoutSettingAs(null);
     });
 }
 
@@ -703,20 +718,26 @@ function renameLayoutSettingAs (evt) {
     div.style.display = 'block';
     evt.stopPropagation();
 	$.get('/result/service/getlayoutsavenames', {'username': username, 'job_id': jobId, 'dbpath': dbPath}).done(function (response) {
-        var quickSaveNameIdx = response.indexOf(quickSaveName);
-        if (quickSaveNameIdx >= 0) {
-            response.splice(quickSaveNameIdx, 1);
+        if (response.length == 0) {
+            var a = getEl('a');
+            a.textContent = '(no\xa0saved\xa0layout)';
+            addEl(div, a);
+        } else {
+            var quickSaveNameIdx = response.indexOf(quickSaveName);
+            if (quickSaveNameIdx >= 0) {
+                response.splice(quickSaveNameIdx, 1);
+            }
+            savedLayoutNames = response;
+            for (var i = 0; i < savedLayoutNames.length; i++) {
+                var name = savedLayoutNames[i];
+                var a = getEl('a');
+                a.textContent = name;
+                a.addEventListener('click', function (evt) {
+                    renameLayoutSetting(evt.target.textContent, null)
+                });
+                addEl(div, a);
+            }
         }
-    	savedLayoutNames = response;
-    	for (var i = 0; i < savedLayoutNames.length; i++) {
-    		var name = savedLayoutNames[i];
-    		var a = getEl('a');
-    		a.textContent = name;
-    		a.addEventListener('click', function (evt) {
-    			renameLayoutSetting(evt.target.textContent, null)
-    		});
-    		addEl(div, a);
-    	}
     });
 }
